Rename resources to content.
