+++ public/content/devicons.json
