+++ public/content/icons.json
